wpt_babel/content-security-policy/securitypolicyviolation/style-sample-no-opt-in.html

<!DOCTYPE html>
<meta content="style-src 'nonce-abc'" http-equiv="Content-Security-Policy"/>
<script nonce="abc" src="/resources/testharness.js">
</script>
<script nonce="abc" src="/resources/testharnessreport.js">
</script>
<body>
 <script nonce="abc">
  "use strict";

function waitForViolation(el) {
  return new Promise(function (resolve) {
    el.addEventListener('securitypolicyviolation', function (e) {
      return resolve(e);
    });
  });
}
async_test(function (t) {
  var s = document.createElement('style');
  s.innerText = "p { omg: yay !important; }";
  waitForViolation(s).then(t.step_func_done(function (e) {
    assert_equals(e.blockedURI, "inline");
    assert_equals(e.sample, "");
  }));
  document.head.append(s);
}, "Inline style blocks should not have a sample.");
async_test(function (t) {
  var p = document.createElement('p');
  p.setAttribute("style", "omg: yay !important;");
  p.innerText = "Yay!";
  waitForViolation(p).then(t.step_func_done(function (e) {
    assert_equals(e.blockedURI, "inline");
    assert_equals(e.sample, "");
  }));
  document.head.append(p);
}, "Inline style attributes should not have a sample.");
 </script>
</body>
